// src/consts/errors.enum.ts
export enum ErrorCodes {
  INTERNAL_SERVER_ERROR = 'API-PM-500',
  COLLECTION_NOT_FOUND = 'API-PM-1',

  // ACCOUNT MONGO
  MONGO_SEARCH_ACCOUNTS_FAILED = 'API-PM-100',
  MONGO_CREATE_ACCOUNT_FAILED = 'API-PM-101',
  MONGO_UPDATE_ACCOUNT_FAILED = 'API-PM-102',
  MONGO_CHECK_ACCOUNT_EXISTED_FAILED = 'API-PM-103',

  // ACCOUNT SERVICE
  SERVICE_SEARCH_ACCOUNTS_FAILED = 'API-PM-200',
  SERVICE_CREATE_ACCOUNT_FAILED = 'API-PM-201',
  SERVICE_UPDATE_ACCOUNT_FAILED = 'API-PM-202',
  SERVICE_ACCOUNT_NOT_EXISTED = 'API-PM-203',
  SERVICE_DELETE_ACCOUNT_FAILED = 'API-PM-204',

  // POSITION MONGO
  MONGO_SEARCH_POSITIONS_FAILED = 'API-PM-300',
  MONGO_CREATE_POSITION_FAILED = 'API-PM-301',
  MONGO_UPDATE_POSITION_FAILED = 'API-PM-302',
  MONGO_CHECK_POSITION_EXISTED_FAILED = 'API-PM-303',

  // POSITION SERVICE
  SERVICE_SEARCH_POSITIONS_FAILED = 'API-PM-400',
  SERVICE_CREATE_POSITION_FAILED = 'API-PM-401',
  SERVICE_UPDATE_POSITION_FAILED = 'API-PM-402',
  SERVICE_POSITION_EXISTED = 'API-PM-403',
  SERVICE_POSITION_NOT_EXISTED = 'API-PM-404',
  SERVICE_DELETE_POSITION_FAILED = 'API-PM-405',

  // ASSET MONGO
  MONGO_SEARCH_ASSETS_FAILED = 'API-PM-500',

  // ASSET SERVICE
  SERVICE_SEARCH_ASSETS_FAILED = 'API-PM-600',

  // ASSET SECTOR MONGO
  MONGO_SEARCH_ASSET_SECTORS_FAILED = 'API-PM-700',

  // ASSET SECTOR SERVICE
  SERVICE_SEARCH_ASSET_SECTORS_FAILED = 'API-PM-800',

  // ASSET PRICE MONGO
  MONGO_SEARCH_ASSET_PRICES_FAILED = 'API-PM-900',

  // ASSET PRICE SERVICE
  SERVICE_SEARCH_ASSET_PRICES_FAILED = 'API-PM-1000',

  // ASSET DIVIDEND MONGO
  MONGO_SEARCH_ASSET_DIVIDENDS_FAILED = 'API-PM-1100',

  // ASSET DIVIDEND SERVICE
  SERVICE_SEARCH_ASSET_DIVIDENDS_FAILED = 'API-PM-1200',

  // ASSET COUNTRY MONGO
  MONGO_SEARCH_ASSET_COUNTRIES_FAILED = 'API-PM-1300',

  // ASSET COUNTRY SERVICE
  SERVICE_SEARCH_ASSET_COUNTRIES_FAILED = 'API-PM-1400'
}

export enum ErrorMessages {
  COLLECTION_NOT_FOUND = 'collection not found',
  INTERNAL_SERVER_ERROR = 'internal server error',

  // ACCOUNT SERVICE
  SERVICE_ACCOUNT_NOT_EXISTED = 'account does not exist',

  // PORTFOLIO SERVICE
  SERVICE_POSITION_EXISTED = 'position has already existed',
  SERVICE_POSITION_NOT_EXISTED = 'position does not exist'
}